Move WebVieSpy to its own file
